[Rename] #15 테스트 코드 패키지 정리
- 접근제한자를 적극 활용하기 위해 테스트코드에서도 프로덕션 코드의 패키지 구조를
따라가려고 함.
